B ijay's branch (#56)
* Before Last one

* Git Ignore

* Workk

* Register NEw User

* Edit user

* Sprint -1 completed expect REST Convention and will be done after learning  API Security

-> Online database connection
->config File
->Sprint 1

* Event Controllers<Music Contoller and FEstival Section

* The festival section

* Dance Thingssssss

* Push First

* BEfore PSotify APIII

* SPotify With WOrking

* Spotify Service

* Country code by IP

* update

* something with artist Table

* Update homecontroller.php

* DanceEventObject

* FEstival Timetable

* Most Of IT WOrking

* Most of it

Timetable classes new methods

* Update Footer.php

* dfdkfn

* NEw Updated NAming Convention

dfdfd

* Update Footer.php

* Router and Admin Controllers

* fgf

fg

* The Festival Performance Can be added

Performance can be added now

* NEw Address Class WIth Db

* NEW updates with delete performance function

* sdksn

* WORKING FILES ARRAY

* Add Artist WORKING IMAGES

df

* Most OOf Things

* Muchh

* APi KEys

* Edit Venue Is Done

* Final Touch

* made Prices and Edit Venues /Delete Performances/Artist/Price VAt

LEft Performance/Artist Update and Making Information Pages

+++ app/services/ArtistService.php
@@ -65,7 +65,7 @@ class ArtistService
     public function addArtist($data, $images): bool
     {
         // checking if the artist already exists in the database or not
-        if($this->artistRepository->artistExistenceInDatabase($data['artistName'],$data['artistDescription'])){
+        if ($this->artistRepository->artistExistenceInDatabase($data['artistName'], $data['artistDescription'])) {
             throw new DatabaseQueryException("Artist With same name and description already exists");
         }
         $newImagesNamesOthers['others'] = $this->getImagesNameByMovingToDirectory($images['others'], $this->imageuploadDirectory);
@@ -97,10 +97,18 @@ class ArtistService
 
     /**
      * @throws uploadFileFailedException
+     * @throws DatabaseQueryException
      */
     public function deleteArist($artistID): bool
     {
+        if ($this->artistRepository->isArtistParticipating($artistID)) {
+            throw new DatabaseQueryException("Artist is participating in event so cannot be deleted");
+        }
         $this->deleteArtistImagesByArtistId($artistID);
+        $artistImagesIds = $this->artistRepository->getAllImagesIdOfArtist($artistID);
+        if (!empty($artistImagesIds)) {
+            $this->deleteArtistImagesFromDb($artistImagesIds); // this will delete images from database
+        }
         return $this->artistRepository->deleteArtist($artistID);
     }
 
@@ -110,13 +118,26 @@ class ArtistService
     private function deleteArtistImagesByArtistId($artistId): void
     {
         $artistImages = $this->artistRepository->getAllImagesNameByArtistId($artistId);
-        if(!empty($artistImages)){
+        if (!empty($artistImages)) {
             $this->deleteImagesFromDirectory($artistImages, $this->imageuploadDirectory);
         }// if file cannot delete then it will throw exception
     }
-    public function isArtistAvailableAtTime($artistId,$date,$time): bool
+
+    public function isArtistAvailableAtTime($artistId, $date, $time): bool
     {
-        return $this->artistRepository->isArtistAvailableAtTime($artistId,$date,$time);
+        return $this->artistRepository->isArtistAvailableAtTime($artistId, $date, $time);
+    }
+
+    /**
+     * @throws DatabaseQueryException
+     */
+    public function deleteArtistImagesFromDb($imageIds): void
+    {
+        foreach ($imageIds as $imageId) {
+            if (!$this->imageService->deleteImage($imageId)) {
+                throw new DatabaseQueryException("Image could not be deleted");
+            };
+        }
     }
 }
